create ACL users control

+++ app/Repositories/Eloquent/ReplyForumRepository.php
@@ -6,6 +6,7 @@ use App\DTO\Replies\CreateReplyDTO;
 use App\Models\ReplyForum;
 use App\Repositories\Contracts\ReplyRepositoryInterface;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use stdClass;
 
 class ReplyForumRepository implements ReplyRepositoryInterface
@@ -34,6 +35,9 @@ class ReplyForumRepository implements ReplyRepositoryInterface
     {
         if(!$reply = $this->model->find($id)){
             return false;
+        }
+        if(Gate::denies('owner', $reply->user->id)){
+            abort('403', 'Not Authorized');
         }
         return (bool) $reply->delete($id);
     }
